Move OnePieceX to app module
