--- cache/httpjajjionepieceexcombrnumero208midia1
@@ -1,240 +1,0 @@
-<html>
- <head>
-  <title>oioi =P</title> 
-  <style>body{margin:0;background:#000}</style> 
-  <link href="./assets/css/adplayerX.css" rel="stylesheet" type="text/css"> 
-  <script type="text/javascript" src="./assets/js/jquery.min.js"></script> 
-  <script type="text/javascript" src="./assets/js/adplayerX.js?v=wr"></script> 
-  <script async src="//pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script> 
-  <meta charset="utf-8"> 
-  <script type="text/javascript">
-
-	$(function(){
-		
-		$("#advideox").adplayer({
-			share:0,
-			volume: 1,
-			keyboard: true,
-			light: false,
-			brightness: true,
-			logo: {
-				src: "assets/opex.png",
-				position: "TopRight",
-				href: "https://one-piecex.com.br",
-				target: "blank",
-				opacity: "0.5",
-				height: "44",
-				width: "89"
-			},
-			// subtitles: "697_srt.srt",
-			quality: {
-				"default": "480p",
-				"480p": "http://klahadorv2.onepieceex.com.br/episodios/online/OpEx_208_online.webm?st=xNdWjVPClaciG4li0VmWIg&e=1499617180",
-				"360p": "http://klahadorv2.onepieceex.com.br/episodios/lq/OpEx_208_LQ.mp4?st=trPrztANgNbHWPTEr4suHg&e=1499617180"			},
-			
-
-			// adroll: {
-				// src: "",
-				// adrolltime: 5,
-				// adlink: "http://onepiecex.com.br/"
-			// },
-			advertising: {
-				"imgroll-1": {
-					type: "adsense",
-					code: '<ins class="adsbygoogle" style="display:inline-block;width:728px;height:90px" data-ad-client="ca-pub-1585462576453676" data-ad-slot="7101566232"></ins>',
-					href: "http://merry.onepieceex.com.br",
-					time: 20
-				},
-				"imgroll-2": {
-					type: "adsense",
-					code: '<ins class="adsbygoogle" style="display:inline-block;width:728px;height:90px" data-ad-client="ca-pub-1585462576453676" data-ad-slot="7101566232"></ins>',
-					href: "http://merry.onepieceex.com.br",
-					time: 600
-				},
-				"imgroll-2": {
-					type: "adsense",
-					code: '<ins class="adsbygoogle" style="display:inline-block;width:728px;height:90px" data-ad-client="ca-pub-8967472176327030" data-ad-slot="9923400102"></ins>',
-					href: "http://merry.onepieceex.com.br",
-					time: 1100
-				},
-			},
-			
-		});
-		
-	});
-
-</script> 
- </head>
- <body>
-  <div style="max-width:853px;height:480px!important;"> 
-   <video id="advideox" width="853" height="480"></video> 
-  </div> 
-  <script> 
-  document.oncontextmenu = function() {return false;};
-
-  window.onload = function(){ 
-    setTimeout(function() { 
-      var ad = document.querySelector("ins.adsbygoogle");
-      // var ad = document.querySelector("div.ad-layer-settings");
-      
-      // If the ad contains no innerHTML, ad blockers are at work
-      if (ad && ad.innerHTML.replace(/\s/g, "").length == 0) {
-        
-        // Since ad blocks hide ads using CSS too
-        ad.style.cssText = 'display:block !important'; 
-        
-        // You can put any text, image or even IFRAME tags here
-        ad.innerHTML = '<div style="background:#FFF;width:100%;height:100%;text-align:center;font-size:2em"><br />Se gosta do que fazemos aqui, considere desativar o adblock!</div>';
-      console.log('adblock');
-      }
-      
-    }, 2000);
-  }; 
-  
-function mudancaQualidade() {console.log('ok =P')}
-
-function mudancaQualidade() {
-    if ($('.ad-layer-settings').height() == 0) {
-        $('.ad-layer-settings').css("height", "35px");
-        $('.ad-layer-settings').css("text-align", "center");
-        $('.ad-layer-settings').html("Para assistir em <strong>melhor qualidade</strong>, desative o AdBlock neste site!");
-        $('.ad-layer-settings').css("color", "#FFF");
-    }
-}
-
-$(document).ready(function(){
-	// Editando
-	// Removendo temporariamente
-	// $("#setagear").hide();
-	// $("#aviso").hide();
-	
-	var testEl = document.createElement( "video" ),
-		mpeg4, h264, ogg, webm;
-	if ( testEl.canPlayType ) {
-		mpeg4 = "" !== testEl.canPlayType( 'video/mp4; codecs="mp4v.20.8"' );
-		h264 = "" !== ( testEl.canPlayType( 'video/mp4; codecs="avc1.42E01E"' )
-			|| testEl.canPlayType( 'video/mp4; codecs="avc1.42E01E, mp4a.40.2"' ) );
-		ogg = "" !== testEl.canPlayType( 'video/ogg; codecs="theora"' );
-		webm = "" !== testEl.canPlayType( 'video/webm; codecs="vp8, vorbis"' );
-	}
-	
-	if (webm == false){
-		$('.ad-gear').trigger('click');
-		$('.ad-quality-menu').trigger('click');
-		$('.ad-quality-option').last().trigger('click');
-		$('.ad-gear').trigger('click').hide();;
-	}
-	
-	var player = document.getElementById("advideox");
-	proximo = false;
-	localStorage['1'] = 208;
-		
-	$(".ad-video").after($('<a class="proximo" target="_top" href="https://one-piecex.com.br/episodios/online/209"><img src="assets/proximo.png"><span>>></span></a>'));
-	$(".ad-video").after($('<div id="aviso">Este episódio está disponível em outras qualidades! <br />Mantenha o Adblock desativado no nosso site e aproveite.</div>'));
-	$(".ad-video").after($('<div id="setagear" class="arrow bounce"></div>'));
-	
-	$(".proximo").hover(function() {
-		$(".proximo span").stop().animate({ width: "95px" }, 500);
-		$(".proximo img").stop().animate({ opacity: "1"},1000);	
-		$(".proximo span").html('Próximo Episódio!');				
-	}, function() {
-		$(".proximo span").html('>>');
-		$(".proximo span").stop().animate({ width: "30px" }, 500);
-		$(".proximo img").stop().animate({ opacity: "0"},1000);
-	});
-	
-
-	player.addEventListener("timeupdate", salvarTempo);
-	player.addEventListener("play", continuar);
-	
-	if ( localStorage.retorno ){
-		console.log('retorno '+localStorage.retorno);
-	}else {
-		localStorage['retorno'] = 1;
-	}
-	
-	localStorage['retorno'] = parseInt(localStorage.retorno)+1;
-	
-			
-	function continuar(){
-		var avisoAtivo = false;
-		if ($('#aviso').is(":visible") && localStorage.retorno < 5 && avisoAtivo == true){
-			$("#aviso").delay(7000).fadeTo("slow", 0);
-			for(i=0;i<7;i++) {
-				$(".ad-gear").fadeTo('slow', 0.5).fadeTo('slow', 1.0);
-			}
-			
-			for(i=0;i<4;i++) {
-				$("#setagear").fadeTo('slow', 0.3).fadeTo('slow', 0.8);
-			}
-			$("#setagear").hide("slow");
-		} else {
-			$("#setagear").hide();
-			$("#aviso").hide();
-		}
-		var player = document.getElementById("advideox");
-		console.log('play'); 
-		
-		// Continua de onde parou
-		if (localStorage.t1208 && player.currentTime == 0) {
-			console.log(localStorage.t1208); 
-			player.currentTime = localStorage.t1208;
-			
-		}
-
-	};
-	function salvarTempo(){
-		var player = document.getElementById("advideox");
-		
-		var tempoAtual = player.currentTime;
-		var tempoAtual = Math.round(tempoAtual);				
-
-		var duracao = player.duration;
-		var duracao = Math.round(duracao);
-		
-		var quaseFim = duracao - duracao * 10 / 100;
-		var quaseFim = Math.round(quaseFim);
-
-		
-		if ( localStorage.t1208 == false ||  tempoAtual != localStorage.t1208 &&  tempoAtual > 1 ) {
-			console.log(localStorage.t1208);
-			
-			localStorage['t1208'] = tempoAtual;
-
-			// Exibir próximo episódio
-			if (tempoAtual > quaseFim && proximo == false){
-				console.log('perto do fim');
-				proximo = true;
-				$(".proximo").show();
-				$(".proximo span").stop().animate({ width: "95px" }, 500);
-				$(".proximo span").html('Próximo Episódio!');
-				$(".proximo img").stop().animate({ opacity: "1"},1000);		
-				
-				setTimeout(function(){
-					$(".proximo span").html('>>');
-					$(".proximo span").stop().animate({ width: "30px" }, 500);
-					$(".proximo img").stop().animate({ opacity: "0"},1000);
-				}, 5000);
-
-				console.log(quaseFim);
-			}
-			
-			// Remover próximo episódio
-			if (tempoAtual < quaseFim && proximo == true){
-				proximo = false;
-				$(".proximo").hide();
-			}
-			
-		 }
-	}
-});
-	
-
-$(mudancaQualidade);
-
-(adsbygoogle = window.adsbygoogle || []).push({});
-(adsbygoogle = window.adsbygoogle || []).push({});
-(adsbygoogle = window.adsbygoogle || []).push({});
-</script> 
- </body>
-</html>
