+++ cache/httpsbf72onepieceexcombrpl20813244422262
@@ -8,7 +8,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge"> 
   <meta name="viewport" content="width=device-width, initial-scale=1"> 
   <meta name="viewport" content="width=device-width, initial-scale=1"> 
-  <link rel="stylesheet" href="/styles.css?123333"> 
+  <link rel="stylesheet" href="https://bf72.onepieceex.com.br/styles.css?123333"> 
   <link rel="icon" type="image/png" href="/favicon.png"> 
   <link href="https://fonts.googleapis.com/css?family=Titillium+Web" rel="stylesheet" type="text/css"> 
   <link href="https://fonts.googleapis.com/css?family=Titillium+Web:400,600" rel="stylesheet" type="text/css"> 
@@ -17,10 +17,10 @@
       <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
       <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
       <![endif]--> 
-  <link href="//bf72.onepieceex.com.br/assets/css/adplayerX.css?v=12082017a" rel="stylesheet" type="text/css"> 
+  <link href="//bf72.onepieceex.com.br/assets/css/adplayerX.css?v=18042018" rel="stylesheet" type="text/css"> 
   <script src="https://code.jquery.com/jquery.js"></script> 
   <script type="text/javascript" src="//bf72.onepieceex.com.br/assets/js/adplayerX.js?v=1308201711h"></script> 
-  <script type="text/javascript">var autoplay = false;var servidor = 'yasopp';var contador = 4;var medida = 100/contador;var ok;var lqp;var block = false;var temProximo = false;var temProximo = true;proximo = false;ativouPlayer = false;</script> 
+  <script type="text/javascript">var autoplay = false;var servidor = 'piiman';var contador = 4;var medida = 100/contador;var ok;var lqp;var block = false;var temProximo = false;var temProximo = true;proximo = false;ativouPlayer = false;</script> 
   <script type="text/javascript">
       		console.log('======================================================================');
       		console.log('Bem vindo ao player da OPEX. Qualquer problema, entre em contato o/');
@@ -29,8 +29,8 @@
       		console.log('======================================================================');
       		console.log('Seu novo anfitrião é -> ' + servidor );
   //     		if (sessionStorage.getItem('servidor') === null ) {
-  //     			var servidor = 'yasopp';
-  //     			sessionStorage.setItem("servidor", 'yasopp');
+  //     			var servidor = 'piiman';
+  //     			sessionStorage.setItem("servidor", 'piiman');
       			// console.log('Seu novo anfitrião é -> ' + sessionStorage.getItem('servidor') );
 		// 	// console.log('Servidor var-> '+ servidor);
 
@@ -40,7 +40,7 @@
 		// 	console.log('Anfitrião já definido. Aos cuidados de -> ' + sessionStorage.servidor);
 		// }
 
-		var codigo = 'https://'+servidor+'.onepieceex.com.br/player.php?code=plays.envbcmA7thEYlgC5%2FB7st6Xn2.TRclxe';
+		var codigo = 'https://'+servidor+'.onepieceex.com.br/player.php?code=plays.e7DZKISfbkto7E7WLrMl84wCUnuCE7VW';
 		// console.log(codigo);
 	</script> 
   <script type="text/javascript">
@@ -159,10 +159,10 @@
 									volume: 1, 
 									keyboard: true,
 									autoplay: false,
-									light: true,
+									light: false,
 									brightness: true,
 									logo: {
-										src: "/assets/opex.png",
+										src: "//bf72.onepieceex.com.br/assets/opex.png",
 										position: "TopRight",
 										href: "https://one-piecex.com.br",
 										target: "blank",
@@ -303,15 +303,15 @@ function salvarTempo(){
 				// $("#ad-v-player").append('<div id="proximo">O próximo episódio começa em <span>'+regressivo+'</span></div>');
 				if ( temProximo == true ){
 					if ($("#proximo").length == 1){
-						$('#ad-v-player #proximo').html('O próximo episódio começa em <span>'+regressivo+'</span> segundos OU <a target="_parent" href="https://one-piece-x.com.br/episodios/online/209"> clique aqui</a>');
+						$('#ad-v-player #proximo').html('O próximo episódio começa em <span>'+regressivo+'</span> segundos OU <a target="_parent" href="https://onepiece-ex.com.br/episodios/online/209"> clique aqui</a>');
 						$('#ad-v-player #proximo').css( "left", "0" );
 					}else{
-						$('#ad-v-player').append('<div id="proximo">O próximo episódio começa em <span>'+regressivo+'</span> segundos OU <a target="_parent" href="https://one-piece-x.com.br/episodios/online/209"> clique aqui</a> </div>');
+						$('#ad-v-player').append('<div id="proximo">O próximo episódio começa em <span>'+regressivo+'</span> segundos OU <a target="_parent" href="https://onepiece-ex.com.br/episodios/online/209"> clique aqui</a> </div>');
 						$('#ad-v-player #proximo').css( "left", "-5000" );
 					}
 					if ( regressivo < 1 )
 					// window.location.href = "http://bf72.onepieceex.com.br/pl/209/1/874035065";
-				top.location.href = "https://one-piece-x.com.br/episodios/online/209";
+				top.location.href = "https://onepiece-ex.com.br/episodios/online/209";
 			}else {
 				if ($("#proximo").length == 1){
 					$('#ad-v-player #proximo').html('O próximo episódio ainda não foi lançado. Assine as notificações, siga no Facebook e aguarde! (Todo Domingo)');
@@ -324,7 +324,7 @@ function salvarTempo(){
 		}
 
 			$( "#ad-v-player #proximo" ).on( "click", function() {
-				window.top.location = "https://one-piece-x.com.br/episodios/online/209";
+				window.top.location = "https://onepiece-ex.com.br/episodios/online/209";
 			});
 
 			// Remover próximo episódio
@@ -346,16 +346,18 @@ function salvarTempo(){
 
 			player.pause();
 			$("#ad-v-player").append('<div id="continuar"><div id="dialog"><h1>Gostaria de continuar de onde parou?</h1><a href="#" id="nao">Não</a><a href="#" id="sim">Sim</a></div></div>');
-			$( "#sim" ).click(function() {
+			$( "#sim" ).click(function(e) {
 				$( "#continuar" ).remove();
 				player.play();
 				player.currentTime = localStorage.tepisodios208;
-
+				e.preventDefault();
 			});
-			$( "#nao" ).click(function() {
+			$( "#nao" ).click(function(e) {
+				e.preventDefault();
 				$( "#continuar" ).remove();
 				localStorage.removeItem('tepisodios208');
 				player.play();
+				e.preventDefault();
 			});
 
 }
