--- cache/httpsyasopponepieceexcombrplayerphpcodeplaysenvbcmA7thEYlgC52FB7st6Xn2TRclxe
@@ -1,9 +1,0 @@
-<html>
- <head></head>
- <body>
-   {"total":2,
-   "LQ":"https:\/\/yasopp.onepieceex.com.br\/episodios\/lq\/OpEx_208_LQ.mp4?st=FdqSS-Frh6KeXiX42YjrPw&amp;e=1514761923",
-   "Online":"https:\/\/yasopp.onepieceex.com.br\/episodios\/online\/OpEx_208_online.webm?st=4zLaEmRj16HdiWg4o0kdCA&amp;e=1514761923",
-   "andamento":100,"code":0,"mensagem":"Tudo pronto."}
- </body>
-</html>
