--- decoders/onepiecex/src/test/kotlin/brunodles/onepiecex/OnePieceXFactoryTest.kt
@@ -1,2482 +1,0 @@
-package brunodles.onepiecex
-
-import brunodles.animewatcher.explorer.Episode
-import brunodles.urlfetcher.UrlFetcher
-import brunodles.animewatcher.testhelper.FactoryChecker
-import com.greghaskins.spectrum.Spectrum
-import org.junit.runner.RunWith
-
-@RunWith(Spectrum::class)
-class OnePieceXFactoryTest {
-
-    companion object {
-
-        val VALID_URLS = arrayOf(
-                "https://one-piece-x.com.br/episodios/online/208/",
-                "http://one-piece-x.com.br/episodios/online/207/",
-                "one-piece-x.com.br/episodios/online/209")
-        val INVALID_URLS = arrayOf(
-                "https://one-piece-x.com.br/episodios/t04/",
-                "one-piece-x.com.br/episodios/online")
-
-        val currentEpisode = Episode(
-                number = 208,
-                animeName = "One Piece",
-                description = "Os piratas do Foxy!! A Davy Back!",
-                link = "https://one-piece-x.com.br/episodios/online/208/",
-                video = "https://yasopp.onepieceex.com.br/episodios/lq/OpEx_208_LQ.mp4?st=FdqSS-Frh6KeXiX42YjrPw&e=1514761923",
-                nextEpisodes = arrayListOf(
-                        Episode(description = "Primeiro Round! Na Donut Race",
-                                number = 209,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/209"),
-                        Episode(description = "Raposa Prateada Foxy! Violenta sabotagem ofensiva!",
-                                number = 210,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/210"),
-                        Episode(description = "O Segundo Evento! Batendo no Groggy Ring!",
-                                number = 211,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/211"),
-                        Episode(description = "Cartas Vermelhas Velozes! Groggy Ring",
-                                number = 212,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/212"),
-                        Episode(description = "O Terceiro Evento! Voltas e Voltas no Roller Racer!",
-                                number = 213,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/213"),
-                        Episode(description = "O Clí­max da Corrida! Entrando no Round Final",
-                                number = 214,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/214"),
-                        Episode(description = "O Grito Servido com Velocidade! O Pirata Dodgeball! (DUPLO)",
-                                number = 215,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/215"),
-                        Episode(description = "Batalha Final no Precipício! A Hora de Daruma-san!",
-                                number = 216,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/216"),
-                        Episode(description = "Combate",
-                                number = 217,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/217"),
-                        Episode(description = "Válvula de Raios Noro Noro VS O Invulnerável Luffy",
-                                number = 218,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/218"),
-                        Episode(description = "Heróica Luta em Combate Feroz! O Decisivo Final!",
-                                number = 219,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/219"),
-                        Episode(description = "Você perdeu, sua Memória foi roubada?! Quem São Vocês?",
-                                number = 220,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/220"),
-                        Episode(description = "O Garoto Misterioso Com o Chifre e a dedução de Robin!",
-                                number = 221,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/221"),
-                        Episode(description = "Recuperação das Memórias! O Desembarque na Ilha dos Piratas!",
-                                number = 222,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/222"),
-                        Episode(description = "Zoro afia seus Caninos! Uma Luta com um Animal Selvagem",
-                                number = 223,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/223"),
-                        Episode(description = "O Contra-ataque Final do Ladrão de Memórias Mostra Sua Verdadeira Natureza!",
-                                number = 224,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/224"),
-                        Episode(description = "O Homem de Orgulho! Foxy, A Raposa Prateada!",
-                                number = 225,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/225"),
-                        Episode(description = "O Invencível Poder Está Perto? E Um Homem Muito Perigoso!",
-                                number = 226,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/226"),
-                        Episode(description = "Almirante da Marinha Aokiji! A Ameaça de Maior Poder (DUPLO)",
-                                number = 227,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/227"),
-                        Episode(description = "Borracha e Gelo na Luta um a um! Luffy VS Aokiji!",
-                                number = 228,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/228"),
-                        Episode(description = "Corra Trem do Oceano! A Cidade das Águas, Water 7!",
-                                number = 229,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/229"),
-                        Episode(description = "Estaleiro Gigante!",
-                                number = 230,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/230"),
-                        Episode(description = "A Família Franky e Iceburg-san!",
-                                number = 231,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/231"),
-                        Episode(description = "Companhia Galley-La! A Magní­fica Doca Um!",
-                                number = 232,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/232"),
-                        Episode(description = "O Sequestro Pirata e o Navio Pirata Só Espera a Morte",
-                                number = 233,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/233"),
-                        Episode(description = "Salvando um Companheiro! O Ataque Surpresa à Casa Do Franky",
-                                number = 234,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/234"),
-                        Episode(description = "A Grande Luta a Luz da Lua! O Navio Pirata Treme em Tristeza!",
-                                number = 235,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/235"),
-                        Episode(description = "Luffy VS Usopp! O Espí­rito dos Homens Colidindo!",
-                                number = 236,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/236"),
-                        Episode(description = "A Cidade da Água Treme! Iceburg Foi Atingido!",
-                                number = 237,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/237"),
-                        Episode(description = "Humano Borracha VS Cyborg da Respiração de Fogo",
-                                number = 238,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/238"),
-                        Episode(description = "Os Criminosos são os Piratas do Chapéu de Palha? Os Guarda-costas de Water 7",
-                                number = 239,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/239"),
-                        Episode(description = "Despedida Eterna? A Mulher Que Atrai a Escuridão, Nico Robin!",
-                                number = 240,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/240"),
-                        Episode(description = "Pegue Robin! Decisão da Tripulação de Chapéu de Palha!",
-                                number = 241,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/241"),
-                        Episode(description = "O Sinal é a Explosão! A CP9 Começa a se Mover!",
-                                number = 242,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/242"),
-                        Episode(description = "CP9 Desmascarada! Seus Chocantes Rostos!",
-                                number = 243,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/243"),
-                        Episode(description = "O Laço Escondido! Iceburg e Franky!",
-                                number = 244,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/244"),
-                        Episode(description = "Volte Robin! Confronto Contra a CP9!",
-                                number = 245,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/245"),
-                        Episode(description = "Aniquilação dos Chapéu de Palha? A Ameaça do Modelo Leopardo!",
-                                number = 246,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/246"),
-                        Episode(description = "O Homem Que é Amado Por Seu Navio! As Lágrimas de Usopp!",
-                                number = 247,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/247"),
-                        Episode(description = "O Passado de Franky! O Dia Que o Trem do Oceano Funcionou",
-                                number = 248,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/248"),
-                        Episode(description = "Conspiração de Spandam! O Dia em Que o Trem do Oceano Balançou",
-                                number = 249,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/249"),
-                        Episode(description = "O Fim do Lendário Homem! O Dia Que o Trem do Oceano Chorou (DUPLO)",
-                                number = 250,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/250"),
-                        Episode(description = "A Verdade por Trás da Traição! A Triste Decisão de Robin!",
-                                number = 251,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/251"),
-                        Episode(description = "O Apito Que Separa os Companheiros! O Trem do Oceano Corre",
-                                number = 252,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/252"),
-                        Episode(description = "Acusação de Sanji! A Batalha do Trem do Oceano na Tempestade!",
-                                number = 253,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/253"),
-                        Episode(description = "O Grito da Alma de Nami! Ressurreição de Luffy, o Chapéu de Palha!",
-                                number = 254,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/254"),
-                        Episode(description = "Outro Trem do Oceano? A Saída de Rocketman!",
-                                number = 255,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/255"),
-                        Episode(description = "O Resgate de Nossos Amigos! Os Punhos Que Carregam a Destruição Para os Inimigos",
-                                number = 256,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/256"),
-                        Episode(description = "Cortando Pela Onda! Zoro e Luffy, o Combo Supremo",
-                                number = 257,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/257"),
-                        Episode(description = "Um Homem Misterioso Aparece?! O Nome Dele é Sogeking!",
-                                number = 258,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/258"),
-                        Episode(description = "A Batalha dos Cozinheiros! Sanji vs Ramen Kenpo!",
-                                number = 259,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/259"),
-                        Episode(description = "Duelo no telhado! Franky vs Nero!",
-                                number = 260,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/260"),
-                        Episode(description = "Impacto Demônio Cortador! Zoro vs Cortador de Navio T-bone!",
-                                number = 261,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/261"),
-                        Episode(description = "Robin Resiste! O Plano Esperto de Sogeking!",
-                                number = 262,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/262"),
-                        Episode(description = "A Ilha do Julgamento! A Vista de Enies Lobby!",
-                                number = 263,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/263"),
-                        Episode(description = "Iniciar Plano de Invasão! O Avanço do Grupo do Chapéu de Palha!",
-                                number = 264,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/264"),
-                        Episode(description = "Investida de Luffy! Grande Decisiva Batalha na Ilha Judiciária!",
-                                number = 265,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/265"),
-                        Episode(description = "Batalha com os Gigantes! Abram a Segunda Porta",
-                                number = 266,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/266"),
-                        Episode(description = "Os Meios de Escapar São Liberados! Voe Através do Céu, Rocketman!",
-                                number = 267,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/267"),
-                        Episode(description = "Alcancem Luffy! A Repentina Guerra dos Chapéu de Palha! (DUPLO)",
-                                number = 268,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/268"),
-                        Episode(description = "Robin Foi Traída! A Intenção do Governo Mundial!",
-                                number = 269,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/269"),
-                        Episode(description = "Nos Devolva Robin! Luffy VS Blueno!",
-                                number = 270,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/270"),
-                        Episode(description = "Não Parem! Levantem o Sinal de um Contra-ataque!",
-                                number = 271,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/271"),
-                        Episode(description = "Luffy à Vista! Reunião na Estação do Tribunal!",
-                                number = 272,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/272"),
-                        Episode(description = "Tudo para proteger meus Nakama! Gear Second em Atividade",
-                                number = 273,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/273"),
-                        Episode(description = "Responda-nos Robin! O Clamor dos Chapéu de Palha!",
-                                number = 274,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/274"),
-                        Episode(description = "O Passado de Robin! A Garota que foi Chamada de Demônio!",
-                                number = 275,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/275"),
-                        Episode(description = "A Mãe e a Filha Predestinadas! O Nome da Mãe é Olvia!",
-                                number = 276,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/276"),
-                        Episode(description = "A Tragédia de Ohara! O Terror do Buster Call (DUPLO)",
-                                number = 277,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/277"),
-                        Episode(description = "Diga Que Você Quer Viver! Nós Somos Companheiros!!",
-                                number = 278,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/278"),
-                        Episode(description = "Pulem No Precipício! Os Sentimentos de Luffy!!",
-                                number = 279,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/279"),
-                        Episode(description = "O Jeito De Viver de Um Homem! As Técnicas Do Zoro e o Sonho De Usopp!",
-                                number = 280,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/280"),
-                        Episode(description = "Lágrimas Que Criaram um Laço Entre Companheiros! O Mapa-mundi de Nami",
-                                number = 281,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/281"),
-                        Episode(description = "Uma Despedida Fortalece um Homem! Sanji e Chopper",
-                                number = 282,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/282"),
-                        Episode(description = "Tudo por um Companheiro!! A Escuridão de Robin!",
-                                number = 283,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/283"),
-                        Episode(description = "Eu Não Vou Entregar as Plantas! A Decisão de Franky!",
-                                number = 284,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/284"),
-                        Episode(description = "Peguem as Cinco Chaves! Os Chapéu de Palha contra CP9!",
-                                number = 285,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/285"),
-                        Episode(description = "O Poder da Akuma no Mi! Kaku e Jabura se Transformam",
-                                number = 286,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/286"),
-                        Episode(description = "Eu Não Chuto Mesmo Que Eu Morra! O Código de Honra de Sanji",
-                                number = 287,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/287"),
-                        Episode(description = "O Erro de Fukurou! A Minha Cola é a Água da Vida!",
-                                number = 288,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/288"),
-                        Episode(description = "A Nova Técnica do Zoro Explode! O Nome da Katana é Sogeking?",
-                                number = 289,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/289"),
-                        Episode(description = "Fora de Controle! Rumble Proibido de Chopper!",
-                                number = 290,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/290"),
-                        Episode(description = "O Retorno de Luffy Oyabun! Um Sonho ou Bilhete Premiado",
-                                number = 291,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/291"),
-                        Episode(description = "A Grande Corrida de Mochi do Castelo! O plano do Nariz Vermelho",
-                                number = 292,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/292"),
-                        Episode(description = "Kalifa a Usuária das Bolhas! Nami se Aproxima da Armadilha de Sabão!",
-                                number = 293,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/293"),
-                        Episode(description = "O Ecoar de Más Notícias! O Buster Call é Acionado!",
-                                number = 294,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/294"),
-                        Episode(description = "5 Namis? O Contra-ataque com Miragens!",
-                                number = 295,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/295"),
-                        Episode(description = "A Determinação de Nami! Atirem no Chopper Furioso!",
-                                number = 296,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/296"),
-                        Episode(description = "Caçador Sanji Aparece? O Lamento do Lobo Mentiroso!",
-                                number = 297,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/297"),
-                        Episode(description = "O Chute Flamejante! O Banquete de Chutes de Sanji!",
-                                number = 298,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/298"),
-                        Episode(description = "O Incrível Duelo",
-                                number = 299,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/299"),
-                        Episode(description = "Zoro, o Deus da Fúria! A Reencarnação de Asura Mostrada Por Seu Espírito!",
-                                number = 300,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/300"),
-                        Episode(description = "Spandam Chocado! Um Herói Se Levanta na Torre da Justiça!",
-                                number = 301,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/301"),
-                        Episode(description = "O Ápice da Batalha Decisiva!",
-                                number = 302,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/302"),
-                        Episode(description = "O Criminoso é Luffy Oyabun? A Busca pela Grande Sakura Perdida!",
-                                number = 303,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/303"),
-                        Episode(description = "Se Eu Não Vencer, Não Posso Proteger Ninguém! O Gear Third é Ativado!",
-                                number = 304,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/304"),
-                        Episode(description = "O Terrível Passado! A Justiça Negra e Rob Lucci!",
-                                number = 305,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/305"),
-                        Episode(description = "Uma Misteriosa Sereia Aparece? No Limiar da Perda de Consciência!",
-                                number = 306,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/306"),
-                        Episode(description = "A Ilha que Afunda Sob Fogo! O Grito de Arrependimento de Franky!",
-                                number = 307,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/307"),
-                        Episode(description = "Esperem Por Ruffy! Confronto Mortal na Ponte da Hesitação!",
-                                number = 308,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/308"),
-                        Episode(description = "Sentimentos Mostrados com os Punhos! O Poder Máximo da Metralhadora de Luffy",
-                                number = 309,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/309"),
-                        Episode(description = "Um Amigo Que Vem Do Mar! O Laço Mais Forte Dos Chapéu De Palha",
-                                number = 310,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/310"),
-                        Episode(description = "Todos na Grande Fuga! O Caminho da Vitória Pertence aos Piratas",
-                                number = 311,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/311"),
-                        Episode(description = "Obrigado, Merry! A Neve Cai no Mar do Adeus",
-                                number = 312,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/312"),
-                        Episode(description = "Nada de Descanso! O Vice-Almirante Que Tem o Punho do Amor",
-                                number = 313,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/313"),
-                        Episode(description = "A Última Linhagem? O Pai de Luffy é Revelado!",
-                                number = 314,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/314"),
-                        Episode(description = "O Nome é Novo Mundo! Além da Grande Rota",
-                                number = 315,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/315"),
-                        Episode(description = "Shanks em Ação! O Começo de Uma Era Selvagem",
-                                number = 316,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/316"),
-                        Episode(description = "A garota que procura pelo Yagara! Grande Busca pela Cidade das Águas!",
-                                number = 317,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/317"),
-                        Episode(description = "A Mãe é Forte! A Ajuda Doméstica do Zoro Trapalhão",
-                                number = 318,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/318"),
-                        Episode(description = "Sanji Chocado! O Velho Misterioso e a Culinária Intensiva",
-                                number = 319,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/319"),
-                        Episode(description = "Todos Finalmente Procurados! O Bando de Mais de 600 Milhões!",
-                                number = 320,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/320"),
-                        Episode(description = "O Rei das Feras Que Irá Cruzar Os Mares! O Impressionante Navio dos Sonhos Terminado!",
-                                number = 321,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/321"),
-                        Episode(description = "Adeus, Meus Seguidores! A Partida de Franky!",
-                                number = 322,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/322"),
-                        Episode(description = "A Partida da Cidade das Águas! A Distinção do Homem Usopp!",
-                                number = 323,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/323"),
-                        Episode(description = "As Recompensas Mundiais! Os Cidadãos Dançam Enquanto o Navio Veleja!",
-                                number = 324,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/324"),
-                        Episode(description = "A Mais Demoníaca Habilidade! Ace e o Ataque da Escuridão",
-                                number = 325,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/325"),
-                        Episode(description = "Bando Pirata Misterioso! O Sunny e uma Perigosa Armadilha",
-                                number = 326,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/326"),
-                        Episode(description = "Emergência no Sunny! O Bradar do Super Sistema de Velocidade!",
-                                number = 327,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/327"),
-                        Episode(description = "O sonho de afundar no novo mundo! O pirata do desespero, Puzzle!",
-                                number = 328,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/328"),
-                        Episode(description = "O ataque dos assassinos! Começa a grande batalha no gelo!",
-                                number = 329,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/329"),
-                        Episode(description = "O grande plano do Bando do Chapéu de Palha! A Alma Pirata movimenta a bandeira!",
-                                number = 330,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/330"),
-                        Episode(description = "À todo vapor! O poder magnético dos gêmeos!",
-                                number = 331,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/331"),
-                        Episode(description = "A mansão do Caos! Nervoso Don e o Bando Aprisionado!",
-                                number = 332,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/332"),
-                        Episode(description = "O Retorno da Fênix! O sonho da Bandeira Pirata em juramento a um amigo!",
-                                number = 333,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/333"),
-                        Episode(description = "A batalha esquenta! Luffy vs Don Ardente!",
-                                number = 334,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/334"),
-                        Episode(description = "A espera do Novo Mundo! Adeus aos corajosos piratas!",
-                                number = 335,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/335"),
-                        Episode(description = "Chopperman Despede-se! Proteja a Estação de TV da Praia",
-                                number = 336,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/336"),
-                        Episode(description = "Aventura no Mar do Demônio! O Misterioso Esqueleto Flutuando na Neblina",
-                                number = 337,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/337"),
-                        Episode(description = "Prazer em Conhecer Pessoas! Os Verdadeiros Sentimentos do Cavalheiro Esqueleto",
-                                number = 338,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/338"),
-                        Episode(description = "Os Calafrios de um Fenômeno Misterioso! O Desembarque em Thriller Bark",
-                                number = 339,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/339"),
-                        Episode(description = "O homem chamado de gênio! Hogback Aparece!",
-                                number = 340,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/340"),
-                        Episode(description = "Nami em Apuros! A Mansão Zumbi e o Homem Invisível",
-                                number = 341,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/341"),
-                        Episode(description = "A Questão Zumbi! O Diabólico Laboratório do Dr. Hogback",
-                                number = 342,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/342"),
-                        Episode(description = "O Nome Dele é Moria! Armadilha do Grande Ladrão de Almas de Piratas",
-                                number = 343,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/343"),
-                        Episode(description = "Festa da Musica dos Zumbis! O Sinal do Ataque Noturno e o Som das Trevas",
-                                number = 344,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/344"),
-                        Episode(description = "Lotado de Animais? O Maravilhoso Jardim de Perona",
-                                number = 345,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/345"),
-                        Episode(description = "O desaparecimento do bando do Chapéu de Palha! O Misterioso espadachim aparece!",
-                                number = 346,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/346"),
-                        Episode(description = "O Último Cavalheiro! O Zumbi Traidor que Protege Nami!",
-                                number = 347,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/347"),
-                        Episode(description = "Vindo do Céu! Aquele Homem é o Espadachim Hanauta!",
-                                number = 348,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/348"),
-                        Episode(description = "Emergência de Luffy! O Destino da Mais Forte Sombra!",
-                                number = 349,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/349"),
-                        Episode(description = "O Guerreiro chamado de Demônio! Hora da Ressurreição de OZ!",
-                                number = 350,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/350"),
-                        Episode(description = "Acordado depois de 500 anos! OZ Revive!",
-                                number = 351,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/351"),
-                        Episode(description = "Convicção que clama pela vida! Brook protege seu Afro!",
-                                number = 352,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/352"),
-                        Episode(description = "A promessa de um homem não morre!! À espera de um amigo num céu distante",
-                                number = 353,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/353"),
-                        Episode(description = "Vamos nos encontrar novamente!! Brooke e a Promessa do Cabo",
-                                number = 354,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/354"),
-                        Episode(description = "Comida, Nami e Sombras!! O contra ataque nervoso de Luffy",
-                                number = 355,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/355"),
-                        Episode(description = "Usopp, o máximo? Deixa os negativos comigo!",
-                                number = 356,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/356"),
-                        Episode(description = "Morte Repentina dos Zumbis Generais!! Oz em Clima de Aventura!!",
-                                number = 357,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/357"),
-                        Episode(description = "O Cavalheiro das Chamas, Sanji!! Dando chute na falsa cerimônia",
-                                number = 358,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/358"),
-                        Episode(description = "Conexão Invisível? O sonho roubado de Sanji",
-                                number = 359,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/359"),
-                        Episode(description = "Herói Salvador!! O inimigo é a Princesa Invencível",
-                                number = 360,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/360"),
-                        Episode(description = "Perona Aterrorizada!! O U de USO é o U de USOPP",
-                                number = 361,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/361"),
-                        Episode(description = "Golpes de espada dançam no telhado!! Conclusão Zoro VS Ryuma",
-                                number = 362,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/362"),
-                        Episode(description = "Chopper Enfurecido!! As práticas médicas demoníacas de Hogback",
-                                number = 363,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/363"),
-                        Episode(description = "Oz ruge!! Apareça Tripulação do Chapéu de Palha",
-                                number = 364,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/364"),
-                        Episode(description = "O Inimigo é Luffy! O Mais Forte Zumbi VS a Tripulação do Chapéu de Palha",
-                                number = 365,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/365"),
-                        Episode(description = "Absalom Nocauteado!! O raio de amizade da Nami!!",
-                                number = 366,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/366"),
-                        Episode(description = "Primeira Queda!! Golpe Final, Junção dos Chapéu de Palha?",
-                                number = 367,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/367"),
-                        Episode(description = "Invasão Silenciosa! O misterioso visitante, Kuma, o tirano!",
-                                number = 368,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/368"),
-                        Episode(description = "Oz mais Moria. Poder e Inteligência, A Combinação Mais Forte",
-                                number = 369,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/369"),
-                        Episode(description = "Estratégia Secreta pra Virar o Jogo. Surge Nightmare Luffy",
-                                number = 370,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/370"),
-                        Episode(description = "O Bando do Chapéu de Palha Aniquilado. Poder Total das Habildades Kage Kage",
-                                number = 371,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/371"),
-                        Episode(description = "A Batalha Pela Superioridade Começa! Luffy VS Luffy",
-                                number = 372,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/372"),
-                        Episode(description = "A conclusão chega! Entregando o golpe final!",
-                                number = 373,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/373"),
-                        Episode(description = "Corpos Desaparecendo! O Sol da Manhã Brilha na Ilha dos Pesadelos!",
-                                number = 374,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/374"),
-                        Episode(description = "A Interminável Crise! A Ordem Para Exterminar a Tripulação do Chapéu de Palha",
-                                number = 375,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/375"),
-                        Episode(description = "Aquilo que Bloqueia Tudo. O Poder Nikyu Nikyu de Kuma",
-                                number = 376,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/376"),
-                        Episode(description = "A Dor dos Meus Companheiros é Minha Dor. Zoro Luta Disposto a Morrer",
-                                number = 377,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/377"),
-                        Episode(description = "A Promessa Naquele Dia Distante. A Canção dos Piratas e a Pequena Baleia",
-                                number = 378,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/378"),
-                        Episode(description = "O Passado de Brook. Triste Despedida Para a Alegre Tripulação",
-                                number = 379,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/379"),
-                        Episode(description = "Saquê do Binks. A Canção Que Liga o Passado e o Presente",
-                                number = 380,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/380"),
-                        Episode(description = "Um Novo Nakama! O Musico “Trovador Brook”",
-                                number = 381,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/381"),
-                        Episode(description = "A Ameaça Noro Noro. O Retorno de Foxy, A Raposa Prateada",
-                                number = 382,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/382"),
-                        Episode(description = "A Grande Caçada ao Tesouro. Colapso! O Spa Island",
-                                number = 383,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/383"),
-                        Episode(description = "O Duro Esforço de Brook. O Difícil Caminho Para Se Tornar um Verdadeiro Nakama?",
-                                number = 384,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/384"),
-                        Episode(description = "O Meio da Grand Line. Chegamos! A Red Line",
-                                number = 385,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/385"),
-                        Episode(description = "Um Rancor Contra os Chapéus de Palha. A Chegada de Duval da Máscara de Ferro",
-                                number = 386,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/386"),
-                        Episode(description = "Reencontro Predestinado! Salvem o Tritão Aprisionado",
-                                number = 387,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/387"),
-                        Episode(description = "Tragédia! A Verdade de Duval Desmascarada",
-                                number = 388,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/388"),
-                        Episode(description = "Explosão! A Arma Super Secreta do Sunny, o Canhão Gaon",
-                                number = 389,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/389"),
-                        Episode(description = "Desembarcando Para Chegar na Ilha dos Tritões. O Arquipélago Sabaody",
-                                number = 390,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/390"),
-                        Episode(description = "Tirania! Os Governantes de Sabaody, os Tenryuubito",
-                                number = 391,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/391"),
-                        Episode(description = "Novos Rivais Reunidos! Os 11 Supernovas",
-                                number = 392,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/392"),
-                        Episode(description = "O Alvo é Caimie!! Os Sequestradores Malvados se Aproximam",
-                                number = 393,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/393"),
-                        Episode(description = "Salvem a Caimie. A Lembrança Sombria Que Ficou no Arquipélago",
-                                number = 394,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/394"),
-                        Episode(description = "Tempo Limite. Começa o Leilão Humano",
-                                number = 395,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/395"),
-                        Episode(description = "Soco Explosivo! Acabem Com o Leilão",
-                                number = 396,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/396"),
-                        Episode(description = "Grande Pânico! Batalha Mortal no Leilão",
-                                number = 397,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/397"),
-                        Episode(description = "Almirante Kizaru se Move! Tumulto no Arquipélago Sabaody",
-                                number = 398,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/398"),
-                        Episode(description = "Superando o Cerco! Marinha VS Três Capitães",
-                                number = 399,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/399"),
-                        Episode(description = "Roger e Rayleigh - O Rei dos Piratas e Seu Braço Direito",
-                                number = 400,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/400"),
-                        Episode(description = "Impossível Fugir!? O Chute na Velocidade da Luz do Almirante Kizaru!!",
-                                number = 401,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/401"),
-                        Episode(description = "Devastador! A Arma de Combate da Marinha, Os Pacifistas",
-                                number = 402,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/402"),
-                        Episode(description = "Mais Um Formidável Inimigo Aparece! Sentomaru Empunha Seu Machado",
-                                number = 403,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/403"),
-                        Episode(description = "Violento Ataque do Almirante Kizaru A Situação Desesperadora do Bando do Chapéu de Palha",
-                                number = 404,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/404"),
-                        Episode(description = "O Desaparecimento dos Nakamas O Último Dia do Bando do Chapéu de Palha",
-                                number = 405,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/405"),
-                        Episode(description = "Episódio Especial Histórico Luffy-Oyabun Aparece Novamente",
-                                number = 406,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/406"),
-                        Episode(description = "Episódio Especial Histórico Acabem Com a Armadilha da Companhia Thriller!",
-                                number = 407,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/407"),
-                        Episode(description = "Desembarque! Ilha Proibida Para Homens, Amazon Lily",
-                                number = 408,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/408"),
-                        Episode(description = "Depressa! Para Onde Estão os Nakamas - Aventura na Ilha das Mulheres",
-                                number = 409,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/409"),
-                        Episode(description = "Todos Apaixonados! Imperatriz Pirata Hancock",
-                                number = 410,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/410"),
-                        Episode(description = "O Segredo Escondido em Suas Costas Luffy Encontra a Princesa Serpente",
-                                number = 411,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/411"),
-                        Episode(description = "O Julgamento Impiedoso! Margareth é Transformada em Pedra!",
-                                number = 412,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/412"),
-                        Episode(description = "Dificílima Batalha de Luffy! O Poder do Haki das Irmãs Serpentes!!",
-                                number = 413,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/413"),
-                        Episode(description = "Batalha de Habilidades ao Máximo!! Gomu Gomu VS Hebi Hebi",
-                                number = 414,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/414"),
-                        Episode(description = "Confissão da Hancock O Passado Repulsivo das Irmãs",
-                                number = 415,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/415"),
-                        Episode(description = "Salvem Ace! O Novo Destino é a Grande Prisão",
-                                number = 416,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/416"),
-                        Episode(description = "O Amor é um Furacão! Hancock Apaixonada",
-                                number = 417,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/417"),
-                        Episode(description = "O Paradeiro dos Nakamas A Ciência do Tempo e a Ilha Mecânica",
-                                number = 418,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/418"),
-                        Episode(description = "O Paradeiro dos Nakamas A Ilha das Aves Gigantes e o Paraíso Cor de Rosa",
-                                number = 419,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/419"),
-                        Episode(description = "O Paradeiro dos Nakamas A Ponte que Conecta Ilhas e as Plantas Comedoras de Gente",
-                                number = 420,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/420"),
-                        Episode(description = "O Paradeiro dos Nakamas A Princesa Negativa e o Diabo!",
-                                number = 421,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/421"),
-                        Episode(description = "Uma Invasão de Vida ou Morte! Invadindo a Prisão Submarina de Impel Down",
-                                number = 422,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/422"),
-                        Episode(description = "Reunião no Inferno?! O Usuáro da Bara Bara no Mi!",
-                                number = 423,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/423"),
-                        Episode(description = "Atravessando! O Inferno Vermelho O Grande Plano de Buggy",
-                                number = 424,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/424"),
-                        Episode(description = "O Homem Mais Forte da Prisão! O Homem-Veneno Magellan Aparece",
-                                number = 425,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/425"),
-                        Episode(description = "Especial Relacionado ao Filme A Ambição do Leão Dourado em Ação",
-                                number = 426,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/426"),
-                        Episode(description = "Especial Relacionado ao Filme Little East Blue é o Alvo",
-                                number = 427,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/427"),
-                        Episode(description = "Especial Relacionado ao Filme O Ataque Furioso dos Piratas Amigo",
-                                number = 428,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/428"),
-                        Episode(description = "Especial Relacionado ao Filme Batalha Decisiva! Luffy VS Largo",
-                                number = 429,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/429"),
-                        Episode(description = "O Shichibukai Preso! Jinbei, o Primeiro Filho do Mar",
-                                number = 430,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/430"),
-                        Episode(description = "A Armadilha do Chefe da Prisão Saldeath Nível 3, o Inferno da Fome",
-                                number = 431,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/431"),
-                        Episode(description = "O Cisne Libertado! Reunião! Bon Clay",
-                                number = 432,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/432"),
-                        Episode(description = "Chefe Magellan se Move O Cerco do Chapéu de Palha Está Completo!",
-                                number = 433,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/433"),
-                        Episode(description = "Concentrando as Forças! Batalha Decisiva no Nível 4, Inferno de Fogo",
-                                number = 434,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/434"),
-                        Episode(description = "Poderoso Magellan! Bon Clay Foge Diante do Inimigo",
-                                number = 435,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/435"),
-                        Episode(description = "Decisão Final! Luffy Arrisca Tudo no Golpe Final",
-                                number = 436,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/436"),
-                        Episode(description = "Porque Somos Amigos! O Resgate Suicida de Bon Clay",
-                                number = 437,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/437"),
-                        Episode(description = "Paraíso no Inferno! Impel Down Nível 5.5",
-                                number = 438,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/438"),
-                        Episode(description = "Começa a Cura de Luffy! A Habilidade Miraculosa de Iva-san",
-                                number = 439,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/439"),
-                        Episode(description = "Acredito em Milagres! O Grito da Alma de Bon Clay",
-                                number = 440,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/440"),
-                        Episode(description = "Ressureição de Luffy! Começa o Plano de Fuga de Iva-san!!",
-                                number = 441,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/441"),
-                        Episode(description = "A Escolta de Ace Começa. Batalha no Último Andar, Nível 6!",
-                                number = 442,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/442"),
-                        Episode(description = "O Último Time Foi Formado Revirando Impel Down!",
-                                number = 443,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/443"),
-                        Episode(description = "Ainda Mais Confusão! A Invasão do Barba Negra Teach!",
-                                number = 444,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/444"),
-                        Episode(description = "Encontro Perigoso! Barba Negra e Shiryu da Chuva",
-                                number = 445,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/445"),
-                        Episode(description = "Não Tem Preço! Hannyabal Sério",
-                                number = 446,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/446"),
-                        Episode(description = "O Jet Pistol da Raiva Luffy VS Barba Negra",
-                                number = 447,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/447"),
-                        Episode(description = "Parem Magellan! Iva-san Solta seu Golpe Secreto",
-                                number = 448,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/448"),
-                        Episode(description = "O Esquema Inteligente de Magellan! Plano de Fuga Frustrado",
-                                number = 449,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/449"),
-                        Episode(description = "O Desespero do Time de Fuga O",
-                                number = 450,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/450"),
-                        Episode(description = "Crie o Milagre Final Rompendo o Portão da Justiça",
-                                number = 451,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/451"),
-                        Episode(description = "O Destino é o Quartel General da Marinha O Navio Parte Para Salvar Ace!",
-                                number = 452,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/452"),
-                        Episode(description = "O Paradeiro dos Nakamas Relatório da Ilha do Céu e os Ciborgues Animais",
-                                number = 453,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/453"),
-                        Episode(description = "O Paradeiro dos Nakamas Pássaros Gigantes e o Confrontro Rosa",
-                                number = 454,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/454"),
-                        Episode(description = "O Paradeiro dos Nakamas O Exército Revolucionário e a Armadilha da Floresta Carnívora!",
-                                number = 455,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/455"),
-                        Episode(description = "O Paradeiro dos Nakamas A Lápide Gigante e as Cuecas do Bem",
-                                number = 456,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/456"),
-                        Episode(description = "Especial Relembrando o Passado Antes de Marinford O Juramento dos Irmãos",
-                                number = 457,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/457"),
-                        Episode(description = "Especial Relembrando o Passado Antes de Marinford Os Três Almirantes",
-                                number = 458,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/458"),
-                        Episode(description = "Chegou a Hora da Batalha! A Formação Mais Poderosa",
-                                number = 459,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/459"),
-                        Episode(description = "Uma Enorme Frota Aparece Invadam! Piratas do Barba Branca",
-                                number = 460,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/460"),
-                        Episode(description = "O Começo da Batalha! O Barba Branca e o Passado de Ace",
-                                number = 461,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/461"),
-                        Episode(description = "O Poder de Destruir o Mundo! A Habilidade da Gura Gura no Mi",
-                                number = 462,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/462"),
-                        Episode(description = "Queimando tudo até virar cinzas!! O poder do almirante Akainu!",
-                                number = 463,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/463"),
-                        Episode(description = "O descendente do demônio! A investida do pequeno Oz Jr.!",
-                                number = 464,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/464"),
-                        Episode(description = "O vencedor será a justiça Em ação! A estratégia de Sengoku!",
-                                number = 465,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/465"),
-                        Episode(description = "A chegada do time do Chapéu De Palha O campo de batalha fica ainda mais intenso",
-                                number = 466,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/466"),
-                        Episode(description = "Vou te salvar mesmo que eu morra Luffy Vs Marinha – A batalha começa",
-                                number = 467,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/467"),
-                        Episode(description = "Batalhas consecutivas! Exército de usuários de Akumas no Mi VS Exército de usuários Akumas no Mi",
-                                number = 468,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/468"),
-                        Episode(description = "O incidente que mudou Kuma O ataque furioso de Iva-san",
-                                number = 469,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/469"),
-                        Episode(description = "O Corte da Espada Negra se Aproxima de Luffy",
-                                number = 470,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/470"),
-                        Episode(description = "A Estratégia de Aniquilação Começa O Poder dos Pacifistas",
-                                number = 471,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/471"),
-                        Episode(description = "O truque de Akainu! O Barba Branca é incriminado",
-                                number = 472,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/472"),
-                        Episode(description = "A operação do cerco começa! Os piratas do Barba Branca são acuados!!",
-                                number = 473,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/473"),
-                        Episode(description = "A ordem para a execução é dada! Atravessem a barreira do cerco",
-                                number = 474,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/474"),
-                        Episode(description = "Entrando na fase final! A jogada do Barba Branca para virar o jogo",
-                                number = 475,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/475"),
-                        Episode(description = "A força bruta de Luffy! Guerra total na Praça Oris!",
-                                number = 476,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/476"),
-                        Episode(description = "O poder que reduz a vida A volta dos hormônios de tensão",
-                                number = 477,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/477"),
-                        Episode(description = "Por Uma Promessa!! Confronto! Luffy e Coby!",
-                                number = 478,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/478"),
-                        Episode(description = "De frente para a plataforma de execução! O caminho para salvar Ace se abre!!",
-                                number = 479,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/479"),
-                        Episode(description = "O caminho que cada um escolheu Luffy vs Garp!",
-                                number = 480,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/480"),
-                        Episode(description = "Ace está livre! A última ordem do Barba Branca como capitão!",
-                                number = 481,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/481"),
-                        Episode(description = "O poder que queima o fogo O golpe final de Akainu",
-                                number = 482,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/482"),
-                        Episode(description = "Procurando por uma resposta Punhos de Fogo Ace morre no campo de batalha",
-                                number = 483,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/483"),
-                        Episode(description = "O Quartel General da Marinha desmorona! A fúria silenciosa do Barba Branca!",
-                                number = 484,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/484"),
-                        Episode(description = "Acertando as contas Barba Branca vs Piratas do Barba Negra",
-                                number = 485,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/485"),
-                        Episode(description = "O Início do Show As intenções malígnas do Barba Negra reveladas",
-                                number = 486,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/486"),
-                        Episode(description = "A persistência de Akainu! O punho de magma ataca Luffy",
-                                number = 487,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/487"),
-                        Episode(description = "Um grito de desespero Um momento de coragem que muda o destino",
-                                number = 488,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/488"),
-                        Episode(description = "Shanks aparece! A guerra dos maiores finalmente chega ao fim!",
-                                number = 489,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/489"),
-                        Episode(description = "Os poderosos rivais se encaram! O começo de uma",
-                                number = 490,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/490"),
-                        Episode(description = "A chegada na Ilha das Mulheres A cruel realidade tortura Luffy",
-                                number = 491,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/491"),
-                        Episode(description = "O mais forte time de luta-livre! Lutem bravamente, Luffy e Toriko!",
-                                number = 492,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/492"),
-                        Episode(description = "Luffy e Ace A história do encontro dos irmãos!",
-                                number = 493,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/493"),
-                        Episode(description = "Sabo aparece! O garoto do Terminal Cinza",
-                                number = 494,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/494"),
-                        Episode(description = "Eu não fujo O resgate suicida de Ace",
-                                number = 495,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/495"),
-                        Episode(description = "Um dia iremos para o mar! O juramento dos copos dos três pirralhos",
-                                number = 496,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/496"),
-                        Episode(description = "Despedindo-se da família Dadan? Terminado! A base secreta",
-                                number = 497,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/497"),
-                        Episode(description = "Luffy, o aprendiz!? O homem que lutou contra o Rei dos Piratas!",
-                                number = 498,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/498"),
-                        Episode(description = "A batalha contra o tigre gigante! O Capitão será aquele que vencer!",
-                                number = 499,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/499"),
-                        Episode(description = "Liberdade roubada! A armadilha dos nobres se aproxima dos três irmãos",
-                                number = 500,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/500"),
-                        Episode(description = "As chamas estão acesas O Terminal Cinza está em crise",
-                                number = 501,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/501"),
-                        Episode(description = "Onde está a liberdade? A triste partida do garoto",
-                                number = 502,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/502"),
-                        Episode(description = "Conto com você! A carta de um irmão",
-                                number = 503,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/503"),
-                        Episode(description = "Para Cumprir a Promessa. Cada Um Segue a Sua Jorna",
-                                number = 504,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/504"),
-                        Episode(description = "Eu quero vê-los! O grito cheio de lágrima de Luffy",
-                                number = 505,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/505"),
-                        Episode(description = "O bando do Chapéu de Palha chocado! A má notícia",
-                                number = 506,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/506"),
-                        Episode(description = "Reunindo-se com o “Rei das Trevas” Rayleigh! Chegou",
-                                number = 507,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/507"),
-                        Episode(description = "Ao nosso capitão A fuga na Ilha do Céu e o incidente na Ilha de Inverno",
-                                number = 508,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/508"),
-                        Episode(description = "Contato! O grande espadachim Mihawk e a luta desesperada do Zoro",
-                                number = 509,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/509"),
-                        Episode(description = "O sofrimento do Sanji A Rainha retorna para o seu reino",
-                                number = 510,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/510"),
-                        Episode(description = "O improvável retorno! Luffy volta para o Quartel General da Marinha!",
-                                number = 511,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/511"),
-                        Episode(description = "Reportem aos nakamas A grande notícia se espalha!",
-                                number = 512,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/512"),
-                        Episode(description = "Os piratas começam a se mover! O assombroso Novo Mundo",
-                                number = 513,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/513"),
-                        Episode(description = "Sobrevivendo ao inferno O duelo de homem do Sanji",
-                                number = 514,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/514"),
-                        Episode(description = "Eu vou ser ainda mais forte! O Juramento de Zoro ao Capitão",
-                                number = 515,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/515"),
-                        Episode(description = "Luffy começa o treinamento No lugar prometido em 2 anos",
-                                number = 516,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/516"),
-                        Episode(description = "Um novo capítulo começa Reunam-se! Bando do Chapéu de Palha",
-                                number = 517,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/517"),
-                        Episode(description = "Cor pronta para explodir! Luffy VS falso Luffy",
-                                number = 518,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/518"),
-                        Episode(description = "A Marinha entra em ação! O alvo é o bando do Chapéu de Palha",
-                                number = 519,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/519"),
-                        Episode(description = "A conclusão do astro A ameaça do Falso Bando do Chapéu de Palha",
-                                number = 520,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/520"),
-                        Episode(description = "Começa a batalha! Mostrem o resultado do treinamento!",
-                                number = 521,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/521"),
-                        Episode(description = "Todos Reunidos Luffy zarpa para o Novo Mundo",
-                                number = 522,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/522"),
-                        Episode(description = "Uma revelação chocante O Homem que protegeu o Sunny Go",
-                                number = 523,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/523"),
-                        Episode(description = "Combate mortal no mar O Demônio das profundezas aparece",
-                                number = 524,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/524"),
-                        Episode(description = "Desastre no fundo do mar O Bando do Chapéu de Palha se perde",
-                                number = 525,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/525"),
-                        Episode(description = "Erupção Vulcânica Submarina Empurrados para a Ilha dos Tritões",
-                                number = 526,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/526"),
-                        Episode(description = "Aterrissando na Ilha dos Tritões! As adoráveis sereias",
-                                number = 527,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/527"),
-                        Episode(description = "Excitação explosiva! A crise de vida do Sanji!",
-                                number = 528,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/528"),
-                        Episode(description = "A extinção da Ilha dos Tritões!? A profecia de Shyarly",
-                                number = 529,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/529"),
-                        Episode(description = "O Rei da Ilha dos Tritões O Deus do mar Neptune!",
-                                number = 530,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/530"),
-                        Episode(description = "O Palácio Ryugu! Trazidos pelo tubarão que eles salvaram",
-                                number = 531,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/531"),
-                        Episode(description = "A Chorona Medrosa! A Princesa Sereia da Torre Koukaku",
-                                number = 532,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/532"),
-                        Episode(description = "Situação de Extrema Urgência! O Palácio Ryugu Foi Ocupado",
-                                number = 533,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/533"),
-                        Episode(description = "Tremor no Palácio Ryugu! O Caso de Sequestro da Shirahoshi",
-                                number = 534,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/534"),
-                        Episode(description = "A Invasão do Hody Começa o Plano de Vingança",
-                                number = 535,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/535"),
-                        Episode(description = "Batalha Decisiva no Palácio Ryugu! Zoro Vs Hody",
-                                number = 536,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/536"),
-                        Episode(description = "Proteja a Shirahoshi! A Perseguição de Decken",
-                                number = 537,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/537"),
-                        Episode(description = "O Bando Derrotado!? Hody Controla o Palácio Ryugu",
-                                number = 538,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/538"),
-                        Episode(description = "Relembrando a Fatalidade! Nami e os Piratas Tritões",
-                                number = 539,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/539"),
-                        Episode(description = "O Herói Libertador de Escravos O Aventureiro Tiger",
-                                number = 540,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/540"),
-                        Episode(description = "Kizaru Aparece! Tiger Segue Para Uma Armadilha!",
-                                number = 541,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/541"),
-                        Episode(description = "Os Times Estão Formados! Salvem o Chopper",
-                                number = 542,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/542"),
-                        Episode(description = "O Fim do Herói A Verdade Chocante de Tiger",
-                                number = 543,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/543"),
-                        Episode(description = "A Separação dos Piratas Jinbe Vs Arlong",
-                                number = 544,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/544"),
-                        Episode(description = "A Ilha dos Tritões Treme! O Naufrágio do Tenryuubito",
-                                number = 545,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/545"),
-                        Episode(description = "Uma Tragédia Inesperada! A Bala Assassina que Bloqueia o Futuro",
-                                number = 546,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/546"),
-                        Episode(description = "De Volta Ao Presente! Hody Começa A Se Mover",
-                                number = 547,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/547"),
-                        Episode(description = "Abalo no Reino A Ordem Para Execução de Neptune",
-                                number = 548,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/548"),
-                        Episode(description = "Diferenças Inesperadas! Luffy Vs Jinbe",
-                                number = 549,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/549"),
-                        Episode(description = "A Transformação de Hody O Verdadeiro Poder da Droga Maligna!",
-                                number = 550,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/550"),
-                        Episode(description = "A Batalha Decisiva Começa A Praça Gyoncorde!",
-                                number = 551,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/551"),
-                        Episode(description = "Confissão Chocante A Verdade Sobre o Assassinato da Otohime",
-                                number = 552,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/552"),
-                        Episode(description = "As Lágrimas da Shirahoshi! Luffy Finalmente Entra em Cena",
-                                number = 553,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/553"),
-                        Episode(description = "O Grande Confronto! O Bando do Chapéu de Palha Vs 100 Mil Inimigos",
-                                number = 554,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/554"),
-                        Episode(description = "Movimentos Explosivos! A Investida de Zoro e Sanji!!",
-                                number = 555,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/555"),
-                        Episode(description = "Primeira Exibição! As Armas Secretas do Sunny-Go!",
-                                number = 556,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/556"),
-                        Episode(description = "O Pirata de Ferro! A Aparição do Franky Shogun!",
-                                number = 557,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/557"),
-                        Episode(description = "Noah Se Aproxima! A Crise da Destruição da Ilha dos Tritões!",
-                                number = 558,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/558"),
-                        Episode(description = "Depressa Luffy! A Situação Desesperadora da Shirahoshi",
-                                number = 559,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/559"),
-                        Episode(description = "Começa o Combate! Luffy Vs Hody!",
-                                number = 560,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/560"),
-                        Episode(description = "A Batalha Real! O Bando Vs Os Novos Piratas Tritões!",
-                                number = 561,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/561"),
-                        Episode(description = "Luffy Derrotado?! A Hora da Vingança do Hody",
-                                number = 562,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/562"),
-                        Episode(description = "A Realidade Chocante! A Verdadeira Identidade do Hody",
-                                number = 563,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/563"),
-                        Episode(description = "Volte a Zero! Uma Súplica Ardente Para Luffy!",
-                                number = 564,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/564"),
-                        Episode(description = "Luffy Ataca com Tudo! A Explosão do Red Hawk",
-                                number = 565,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/565"),
-                        Episode(description = "O Último Acerto de Contas! A Batalha Final Contra Hody",
-                                number = 566,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/566"),
-                        Episode(description = "Pare Noah! O Elephant Gatling Gun Desesperado!",
-                                number = 567,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/567"),
-                        Episode(description = "Para o Futuro! O Caminho Que Leva Ao Sol!",
-                                number = 568,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/568"),
-                        Episode(description = "O Segredo Revelado A Verdade Sobre as Armas Antigas",
-                                number = 569,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/569"),
-                        Episode(description = "O Bando Surpreso! O Novo Almirante da Frota da Marinha!",
-                                number = 570,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/570"),
-                        Episode(description = "A Adoradora de Doces! A Yonkou Big Mom",
-                                number = 571,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/571"),
-                        Episode(description = "Mares Tempestuosos Uma Armadilha Aguarda no Novo Mundo",
-                                number = 572,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/572"),
-                        Episode(description = "Finalmente Zarpar! Adeus, Ilha dos Tritões",
-                                number = 573,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/573"),
-                        Episode(description = "Para o Novo Mundo! De Olhos para o Mar Mais Forte",
-                                number = 574,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/574"),
-                        Episode(description = "A Ambição do Z Lily, A Pequena Gigante!",
-                                number = 575,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/575"),
-                        Episode(description = "A Ambição de Z Aparece o Mais Forte Exército Misterioso",
-                                number = 576,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/576"),
-                        Episode(description = "A Ambição de Z O Plano da Grande Fuga Desesperada!",
-                                number = 577,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/577"),
-                        Episode(description = "A Ambição de Z Luffy Vs Shuzo",
-                                number = 578,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/578"),
-                        Episode(description = "Desembarque! Punk Hazard, A Ilha Flamejante",
-                                number = 579,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/579"),
-                        Episode(description = "Uma Batalha Muito Quente! Luffy Vs Dragão Gigante",
-                                number = 580,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/580"),
-                        Episode(description = "O Bando Confuso! Aparece o Chocante Samurai Decapitado!",
-                                number = 581,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/581"),
-                        Episode(description = "Surpresa! O Segredo da Ilha é Finalmente Revelado",
-                                number = 582,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/582"),
-                        Episode(description = "Salvem as Crianças! Começa a Batalha do Bando!",
-                                number = 583,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/583"),
-                        Episode(description = "Um Duelo de Habilidade de Espadas Brook Vs O Corpo do Samurai Misterioso",
-                                number = 584,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/584"),
-                        Episode(description = "Shichibukai! Trafalgar Law",
-                                number = 585,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/585"),
-                        Episode(description = "O Grande Aperto Luffy Afunda no Lago Congelado",
-                                number = 586,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/586"),
-                        Episode(description = "Colisão Violenta! Law Vs. Vice-Almirante Smoker",
-                                number = 587,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/587"),
-                        Episode(description = "Primeiro Encontro em 2 Anos! Luffy e Law",
-                                number = 588,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/588"),
-                        Episode(description = "O Pior do Mundo Caesar, o Cientista Terrível",
-                                number = 589,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/589"),
-                        Episode(description = "A Mais Forte Colaboração da História Vs O Glutão dos Mares",
-                                number = 590,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/590"),
-                        Episode(description = "Chopper Furioso Os Experimentos Desumanos do Mestre",
-                                number = 591,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/591"),
-                        Episode(description = "O Bando Dizimado! O Ataque dos Assassinos Lendários!",
-                                number = 592,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/592"),
-                        Episode(description = "Salvem a Nami! A Batalha de Luffy nas Montanhas de Neve",
-                                number = 593,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/593"),
-                        Episode(description = "Unidos! A Aliança Pirata de Luffy e Law!",
-                                number = 594,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/594"),
-                        Episode(description = "Capturem o Mestre Começa a Estratégia da Aliança Pirata!",
-                                number = 595,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/595"),
-                        Episode(description = "Crise Destrutiva O Monstro da Morte Vem Voando",
-                                number = 596,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/596"),
-                        Episode(description = "A Grande Batalha Feroz Caesar Ativa Sua Verdadeira Habilidade!",
-                                number = 597,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/597"),
-                        Episode(description = "O Samurai que Corta Fogo! Kinemon, o Raposa de Fogo!",
-                                number = 598,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/598"),
-                        Episode(description = "Surpresa! A Verdadeira Identidade de Vergo, o Homem Misterioso!",
-                                number = 599,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/599"),
-                        Episode(description = "Protejam as Crianças! Aproxima-se a Mão Maligna do Mestre",
-                                number = 600,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/600"),
-                        Episode(description = "Tremor no Novo Mundo O Experimento Pesadelo de Caesar",
-                                number = 601,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/601"),
-                        Episode(description = "A Pior Arma Assassina da História! Shinokuni",
-                                number = 602,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/602"),
-                        Episode(description = "Começa o Contra-Ataque! A Grande Escapada de Luffy e Law",
-                                number = 603,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/603"),
-                        Episode(description = "Em Direção ao Setor-R! A Grande Invasão da Aliança Pirata",
-                                number = 604,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/604"),
-                        Episode(description = "As Lágrimas da Tashigi A Estratégia de Avanço Suicida da G5",
-                                number = 605,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/605"),
-                        Episode(description = "A Traição do Vice-Almirante! Vergo, do Bambu Demoníaco",
-                                number = 606,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/606"),
-                        Episode(description = "Uma Feroz Batalha se Aquece! Luffy Vs. Caesar!",
-                                number = 607,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/607"),
-                        Episode(description = "O Mentor das Sombras! Doflamingo se Move!",
-                                number = 608,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/608"),
-                        Episode(description = "Luffy Congelando até a Morte?! A Terrível Mulher-Neve, Monet!",
-                                number = 609,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/609"),
-                        Episode(description = "Colisão de Punhos! A Batalha de Dois Vice-Almirantes",
-                                number = 610,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/610"),
-                        Episode(description = "O Pequeno Dragão! Mononosuke Aparece",
-                                number = 611,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/611"),
-                        Episode(description = "Batalha Mortal na Tempestade de Neve Os Chapéus de Palha Vs A Mulher-Neve",
-                                number = 612,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/612"),
-                        Episode(description = "Explosão Misteriosa! O Ittoryu Mais Forte de Zoro!",
-                                number = 613,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/613"),
-                        Episode(description = "Vou Proteger Meus Amigos! Mocha Corre Arriscando Sua Vida",
-                                number = 614,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/614"),
-                        Episode(description = "A Amargura do Barba Marrom! O Ataque Furioso de Luffy",
-                                number = 615,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/615"),
-                        Episode(description = "Conclusão Surpreendente! Smoker Vs. Vergo!",
-                                number = 616,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/616"),
-                        Episode(description = "Caesar é Derrotado! O Poderoso Grizzly Magnum",
-                                number = 617,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/617"),
-                        Episode(description = "Invasão! Os Assassinos de Dressrosa",
-                                number = 618,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/618"),
-                        Episode(description = "O Grande Alvoroço! O Invencível Franky Shogun",
-                                number = 619,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/619"),
-                        Episode(description = "Morte Absoluta! A Grande Explosão de Punk Hazard",
-                                number = 620,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/620"),
-                        Episode(description = "Capturem Caesar A Explosão do General Cannon",
-                                number = 621,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/621"),
-                        Episode(description = "Reencontro Emocionante! Momonosuke e Kinemon",
-                                number = 622,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/622"),
-                        Episode(description = "Hora de uma Despedida Triste Partindo de Punk Hazard!",
-                                number = 623,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/623"),
-                        Episode(description = "O Extermínio da G5! O Ataque Surpresa de Doflamingo!",
-                                number = 624,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/624"),
-                        Episode(description = "Tensão! Aokiji Vs Doflamingo",
-                                number = 625,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/625"),
-                        Episode(description = "Caesar Desaparece! A Investida da Aliança Pirata",
-                                number = 626,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/626"),
-                        Episode(description = "Luffy Morre no Mar?! O Colapso da Aliança Pirata",
-                                number = 627,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/627"),
-                        Episode(description = "Grande Reviravolta! A Explosão da Fúria do Punho de Luffy",
-                                number = 628,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/628"),
-                        Episode(description = "Impacto Violento! A Grande Notícia que Abala o Mundo",
-                                number = 629,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/629"),
-                        Episode(description = "Aventura! Dressrosa, o País do Amor e da Paixão",
-                                number = 630,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/630"),
-                        Episode(description = "Um Redemoinho de Entusiasmos O Coliseu Corrida",
-                                number = 631,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/631"),
-                        Episode(description = "Paixão Perigosa A Dançarina Violet",
-                                number = 632,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/632"),
-                        Episode(description = "O Guerreiro Desconhecido Mais Poderoso! Lucy Aparece",
-                                number = 633,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/633"),
-                        Episode(description = "O Príncipe Pirata Cavendish",
-                                number = 634,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/634"),
-                        Episode(description = "Reencontro do Destino Bellamy, o Hiena",
-                                number = 635,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/635"),
-                        Episode(description = "Supernova! Bartolomeo, o Canibal",
-                                number = 636,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/636"),
-                        Episode(description = "A Rivalidade dos Principais Guerreiros! O Ápice do Bloco B!",
-                                number = 637,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/637"),
-                        Episode(description = "Nocautes de Um Golpe Só! O Assombroso King Punch",
-                                number = 638,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/638"),
-                        Episode(description = "O Ataque dos Peixes-Lutadores! Passem Pela Ponte de Ferro da Morte",
-                                number = 639,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/639"),
-                        Episode(description = "Aventura! A Ilha das Fadas, Green Bit",
-                                number = 640,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/640"),
-                        Episode(description = "O Mundo Desconhecido O Reino de Tontatta",
-                                number = 641,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/641"),
-                        Episode(description = "O Truque do Século Doflamingo em Ação",
-                                number = 642,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/642"),
-                        Episode(description = "Céus e Terra Estremecem! O Verdadeiro Poder do Almirante Fujitora",
-                                number = 643,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/643"),
-                        Episode(description = "Um Golpe de Raiva! Gigante Vs Lucy",
-                                number = 644,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/644"),
-                        Episode(description = "O Canhão Demolidor Explode! Lucy em Apuros",
-                                number = 645,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/645"),
-                        Episode(description = "O Pirata Lendário. Don Chinjao!",
-                                number = 646,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/646"),
-                        Episode(description = "Luz e Sombras A Escuridão Oculta de Dressrosa!",
-                                number = 647,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/647"),
-                        Episode(description = "Avante Herói Lendário Usoland",
-                                number = 648,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/648"),
-                        Episode(description = "Conclusão da Batalha Feroz Lucy VS Chinjao",
-                                number = 649,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/649"),
-                        Episode(description = "Luffy e a Gladiadora Predestinada Rebecca",
-                                number = 650,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/650"),
-                        Episode(description = "Vou Protegê-la Até o Fim! Rebecca e o Soldado de Brinquedo",
-                                number = 651,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/651"),
-                        Episode(description = "O Último Vale-Tudo. O Início da Batalha no Bloco D.",
-                                number = 652,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/652"),
-                        Episode(description = "A Batalha Decisiva! Jora Vs Bando dos Chapéus de Palha",
-                                number = 653,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/653"),
-                        Episode(description = "A Espada Linda! Cavendish do Cavalo Branco!",
-                                number = 654,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/654"),
-                        Episode(description = "Grande Impacto! Sanji Vs Doflamingo",
-                                number = 655,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/655"),
-                        Episode(description = "A Espada Mortal de Rebecca! A Dança das Espadas de Água",
-                                number = 656,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/656"),
-                        Episode(description = "O Guerreiro Mais Malvado! Logan VS Rebecca",
-                                number = 657,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/657"),
-                        Episode(description = "Surpresa! A Verdadeira Identidade do Soldado de Brinquedo",
-                                number = 658,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/658"),
-                        Episode(description = "Um Passado de Arrepiar! O Segredo de Dressrosa",
-                                number = 659,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/659"),
-                        Episode(description = "Pesadelo! Aquela Noite Trágica de Dressrosa",
-                                number = 660,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/660"),
-                        Episode(description = "Confronto de Shichibukais Law Vs Doflamingo",
-                                number = 661,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/661"),
-                        Episode(description = "Dois Grandes Homens se Encaram! Chapéu de Palha e Demônio Celestial",
-                                number = 662,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/662"),
-                        Episode(description = "Luffy Espantado O Homem que Herdou a Determinação de Ace",
-                                number = 663,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/663"),
-                        Episode(description = "A Operação SOP Começa O Ataque de Usoland",
-                                number = 664,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/664"),
-                        Episode(description = "Fortes Emoções Rebecca Vs Suleiman",
-                                number = 665,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/665"),
-                        Episode(description = "Vencedor Decidido?! O Desfecho Devastador do Bloco D",
-                                number = 666,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/666"),
-                        Episode(description = "A Decisão do Almirante Fujitora Vs Doflamingo",
-                                number = 667,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/667"),
-                        Episode(description = "Começa a Final O Herói Dia Mante Aparece",
-                                number = 668,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/668"),
-                        Episode(description = "O Castelo se Mexe! A Aparição do Comandante Pica",
-                                number = 669,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/669"),
-                        Episode(description = "A Garra do Dragão Explode! O Golpe Assombroso de Lucy!",
-                                number = 670,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/670"),
-                        Episode(description = "Derrotem Sugar. O Ataque dos Soldados Anões!",
-                                number = 671,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/671"),
-                        Episode(description = "A Última Luz O Segredo do Nosso Comandante",
-                                number = 672,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/672"),
-                        Episode(description = "O Homem Estouro. As Grandes Explosões de Gladius!",
-                                number = 673,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/673"),
-                        Episode(description = "O Mentiroso. Usoland em Fuga!",
-                                number = 674,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/674"),
-                        Episode(description = "O Encontro do Destino. Kyros e Rei Riku.",
-                                number = 675,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/675"),
-                        Episode(description = "A Operação Fracassou! O Herói Usoland Morreu?!",
-                                number = 676,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/676"),
-                        Episode(description = "A Lenda Revive! Kyros Ataca com Tudo.",
-                                number = 677,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/677"),
-                        Episode(description = "A Explosão dos Punhos de Fogo! Renasce o Poder da Mera Mera no Mi.",
-                                number = 678,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/678"),
-                        Episode(description = "Uma Impetuosa Aparição. Sabo, o Conselheiro do Líder do Exército Revolucionário!",
-                                number = 679,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/679"),
-                        Episode(description = "A Armadilha do Demônio. O Plano de Extermínio de Dressrosa",
-                                number = 680,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/680"),
-                        Episode(description = "O Homem de 500 Milhões – O Alvo é Usoland!",
-                                number = 681,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/681"),
-                        Episode(description = "Rompendo as Frentes Inimigas Começa o Contra-Ataque de Luffy e Zoro",
-                                number = 682,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/682"),
-                        Episode(description = "A Terra Treme A Epifania do Deus Colossal da Destruição Pica",
-                                number = 683,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/683"),
-                        Episode(description = "Uma Poderosa Linha de Frente! Luffy e o Exército dos Guerreiros Brutais",
-                                number = 684,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/684"),
-                        Episode(description = "A Grande Ofensiva! O Exército de Luffy Vs Pica!",
-                                number = 685,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/685"),
-                        Episode(description = "Uma Confissão Chocante! O Juramento Intenso de Law!",
-                                number = 686,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/686"),
-                        Episode(description = "O Grande Confronto! Conselheiro-Líder Sabo Vs Almirante Fujitora",
-                                number = 687,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/687"),
-                        Episode(description = "Uma Situação Desesperadora. Luffy Cai em uma Armadilha!",
-                                number = 688,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/688"),
-                        Episode(description = "A Grande Escapada! O Milagroso Elephant Gun de Luffy.",
-                                number = 689,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/689"),
-                        Episode(description = "A Frente de Batalha Unida O Avanço de Luffy para a Vitória",
-                                number = 690,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/690"),
-                        Episode(description = "O Segundo Samurai. Kanjurou do Chuvisco Aparece.",
-                                number = 691,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/691"),
-                        Episode(description = "O Embate Violento com Pica O Golpe Mortal de Zoro!",
-                                number = 692,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/692"),
-                        Episode(description = "A Princesa dos Anões. A Prisioneira Mansherry.",
-                                number = 693,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/693"),
-                        Episode(description = "Invencíveis! O Terrível Exército dos Quebra-Nozes.",
-                                number = 694,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/694"),
-                        Episode(description = "Arriscando Nossas Vidas! Luffy é o Trunfo para a Vitória.",
-                                number = 695,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/695"),
-                        Episode(description = "Um Reencontro de Lágrimas Rebecca e Kyros!",
-                                number = 696,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/696"),
-                        Episode(description = "O Tiro Certeiro O Homem que Salvará Dressrosa",
-                                number = 697,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/697"),
-                        Episode(description = "Explosão de Raiva O Poderoso Plano Secreto de Luffy e Law",
-                                number = 698,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/698"),
-                        Episode(description = "Uma Nobre Família A Verdadeira Identidade de Doflamingo!",
-                                number = 699,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/699"),
-                        Episode(description = "O Poder Supremo O Segredo da Ope Ope no Mi!",
-                                number = 700,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/700"),
-                        Episode(description = "Memórias Tristes. Law, o Menino da Cidade Branca!",
-                                number = 701,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/701"),
-                        Episode(description = "O Tenryuubito! O Passado Tempestuoso de Doffy",
-                                number = 702,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/702"),
-                        Episode(description = "Um Caminho de Sofrimento. A Jornada Pela Vida de Law e Corazón.",
-                                number = 703,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/703"),
-                        Episode(description = "O Tempo está Acabando! A Captura da Ope Ope no Mi!",
-                                number = 704,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/704"),
-                        Episode(description = "Momento da Decisão. O Sorriso de Despedida de Corazón!",
-                                number = 705,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/705"),
-                        Episode(description = "Vá, Law. A Batalha Final de um Homem de Bom Coração!",
-                                number = 706,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/706"),
-                        Episode(description = "Pela Liberdade! Law Dispara o Injection Shot",
-                                number = 707,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/707"),
-                        Episode(description = "Uma Batalha Intensa. Law Vs Doflamingo.",
-                                number = 708,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/708"),
-                        Episode(description = "A Batalha Decisiva Contra os Oficiais. O Grande Orgulho de Hajrudin.",
-                                number = 709,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/709"),
-                        Episode(description = "A Batalha do Amor. O Novo Líder Sai VS Baby 5.",
-                                number = 710,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/710"),
-                        Episode(description = "O Orgulho de um Homem. O Último Ataque de Bellamy!",
-                                number = 711,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/711"),
-                        Episode(description = "Vendaval Tempestuoso. Hakuba VS Dellinger.",
-                                number = 712,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/712"),
-                        Episode(description = "Bari Bari. A Homenagem do Punho Divino Ativado!",
-                                number = 713,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/713"),
-                        Episode(description = "A Princesa da Cura. Salvem Mansherry.",
-                                number = 714,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/714"),
-                        Episode(description = "O Duelo Entre Homens. O Réquiem do Amor do Señor.",
-                                number = 715,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/715"),
-                        Episode(description = "Poeira Estelar da Morte. A Tempestade de Ataques Violentos de Diamante.",
-                                number = 716,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/716"),
-                        Episode(description = "Trueno Bastardo! O Golpe de Fúria de Kyros!",
-                                number = 717,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/717"),
-                        Episode(description = "Se Movendo pelo Chão. A Estratégia do Ataque-Surpresa do Gigantesco Pica!",
-                                number = 718,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/718"),
-                        Episode(description = "Batalha Final no Ar. A Explosão da Nova Técnica Secreta Mortal do Zoro!",
-                                number = 719,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/719"),
-                        Episode(description = "Até Mais! O Golpe de Adeus do Bellamy!",
-                                number = 720,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/720"),
-                        Episode(description = "Law Morre. O Ataque Violento de Raiva do Luffy!",
-                                number = 721,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/721"),
-                        Episode(description = "A Lâmina da Determinação O Contra-Ataque da Gamma Knife!",
-                                number = 722,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/722"),
-                        Episode(description = "Colisão de Haki. Luffy Vs Doflamingo.",
-                                number = 723,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/723"),
-                        Episode(description = "Impossível Atacar. O Segredo Chocante de Trebol.",
-                                number = 724,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/724"),
-                        Episode(description = "Explosão de Raiva. Eu me Responsabilizo por Tudo.",
-                                number = 725,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/725"),
-                        Episode(description = "Gear 4! O Fenomenal Boundman!",
-                                number = 726,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/726"),
-                        Episode(description = "O Grande Contra-Ataque! O Despertar de Doflamingo!",
-                                number = 727,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/727"),
-                        Episode(description = "Luffy! O Devastador Leo Bazooka!",
-                                number = 728,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/728"),
-                        Episode(description = "A Chama do Rei Dragão. Protejam a Vida de Luffy.",
-                                number = 729,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/729"),
-                        Episode(description = "Lágrimas Milagrosas. A Luta de Mansherry!",
-                                number = 730,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/730"),
-                        Episode(description = "Enquanto Estivermos Vivos. Parem a Gaiola da Morte!",
-                                number = 731,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/731"),
-                        Episode(description = "Vida ou Morte. A Contagem Regressiva do Destino.",
-                                number = 732,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/732"),
-                        Episode(description = "Derrubando os Céus. A Fúria do King Kong Gun de Luffy.",
-                                number = 733,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/733"),
-                        Episode(description = "Para a Liberdade! Alegre-se, Dressrosa!",
-                                number = 734,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/734"),
-                        Episode(description = "Sem Precedentes. A Impactante Decisão do Almirante Fujitora!",
-                                number = 735,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/735"),
-                        Episode(description = "Sacudindo o Mundo. A Pior Geração Avança.",
-                                number = 736,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/736"),
-                        Episode(description = "O Nascimento de uma Lenda. A Aventura do Revolucionário Sabo!",
-                                number = 737,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/737"),
-                        Episode(description = "O Laço dos Irmãos. A História Secreta do Reencontro de Luffy e Sabo.",
-                                number = 738,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/738"),
-                        Episode(description = "A Criatura Mais Forte. O Yonkou Kaidou, o Rei das Feras.",
-                                number = 739,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/739"),
-                        Episode(description = "Fujitora se Move. O Bando dos Chapéus de Palha Completamente Cercado.",
-                                number = 740,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/740"),
-                        Episode(description = "Estado de Emergência. Rebecca é Sequestrada!",
-                                number = 741,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/741"),
-                        Episode(description = "Laços de Pai e Filha. Kyros e Rebecca!",
-                                number = 742,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/742"),
-                        Episode(description = "A Determinação de um Homem. Luffy Vs Fujitora Cara a Cara.",
-                                number = 743,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/743"),
-                        Episode(description = "Sem saída. A Perseguição Implacável do Almirante Fujitora!",
-                                number = 744,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/744"),
-                        Episode(description = "Os Seguidores dos Copos de Saquê. Está Formada! A Grande Frota do Chapéu de Palha!",
-                                number = 745,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/745"),
-                        Episode(description = "A Fúria Incontrolável dos Poderosos. Os Monstros do Novo Mundo.",
-                                number = 746,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/746"),
-                        Episode(description = "A Fortaleza Prateada. A Grande Aventura de Luffy e Barto.",
-                                number = 747,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/747"),
-                        Episode(description = "O Labirinto Subterrâneo. Luffy Vs O Homem-Bonde.",
-                                number = 748,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/748"),
-                        Episode(description = "A Técnica de Espadas Ferve. Law e Zoro Finalmente Aparecem!",
-                                number = 749,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/749"),
-                        Episode(description = "Situação Desesperadora. A Batalha Escaldante de Luffy.",
-                                number = 750,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/750"),
-                        Episode(description = "O Início da Aventura. A Chegada na Ilha Mística de Zou!",
-                                number = 751,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/751"),
-                        Episode(description = "O Novo Shichibukai. O Filho do Lendário Barba Branca Aparece.",
-                                number = 752,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/752"),
-                        Episode(description = "A Escalada Suicida no elefante. Uma Grande Aventura nas Costas do Elefante Gigante!",
-                                number = 753,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/753"),
-                        Episode(description = "Começa o Combate. Luffy Vs Tribo Mink!",
-                                number = 754,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/754"),
-                        Episode(description = "Garchu! Os Chapéus de Palha Reunidos.",
-                                number = 755,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/755"),
-                        Episode(description = "Começa o Contra-Ataque. As Grandes Táticas do Bando do Sobrancelha Encaracolada!",
-                                number = 756,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/756"),
-                        Episode(description = "A Invasão Ameaçadora. Jack dos Piratas das Feras.",
-                                number = 757,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/757"),
-                        Episode(description = "O Rei do Dia. Duque Inuarashi Aparece!",
-                                number = 758,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/758"),
-                        Episode(description = "O Rei da Noite. Surge o Mestre Nekomamushi.",
-                                number = 759,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/759"),
-                        Episode(description = "A Destruição da Capital. O Desembarque do Bando do Sobrancelha Encaracolada!",
-                                number = 760,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/760"),
-                        Episode(description = "Corrida Contra o Tempo. O Laço da Tribo Mink com o Bando!",
-                                number = 761,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/761"),
-                        Episode(description = "O Delinquente Volta para Casa. Os Assassinos da Yonkou Big Mom.",
-                                number = 762,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/762"),
-                        Episode(description = "A verdade por trás do desaparecimento. Sanji recebe um convite espantoso.",
-                                number = 763,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/763"),
-                        Episode(description = "Caros Amigos de Merda. O Recado de Despedida do Sanji.",
-                                number = 764,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/764"),
-                        Episode(description = "Vamos Conhecer o Mestre Nekomamushi.",
-                                number = 765,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/765"),
-                        Episode(description = "A Decisão de Luffy. A Crise do Afastamento do Sanji!",
-                                number = 766,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/766"),
-                        Episode(description = "Situação Extremamente Perigosa. O Cão, o Gato e Samurais!",
-                                number = 767,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/767"),
-                        Episode(description = "O Terceiro! O Ninja Raizou da Névoa Aparece.",
-                                number = 768,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/768"),
-                        Episode(description = "A Pedra Vermelha! O Guia para o One Piece.",
-                                number = 769,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/769"),
-                        Episode(description = "O Segredo do País de Wano. O Clã Kouzuki e os Poneglyphs.",
-                                number = 770,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/770"),
-                        Episode(description = "A Promessa de Homens. Luffy e Kouzuki Momonosuke.",
-                                number = 771,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/771"),
-                        Episode(description = "A Jornada Lendária. O Cão, o Gato e Rei dos Piratas!",
-                                number = 772,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/772"),
-                        Episode(description = "O Retorno do Pesadelo. O Ataque Violento do Indestrutível Jack.",
-                                number = 773,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/773"),
-                        Episode(description = "A Batalha pela Defesa de Zou. Luffy e Zunisha!",
-                                number = 774,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/774"),
-                        Episode(description = "Salvem o Elefante Gigante. O Plano de Resgate dos Chapéus de Palha!",
-                                number = 775,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/775"),
-                        Episode(description = "Descendo do Elefante. Zarpando para Resgatar o Sanji!",
-                                number = 776,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/776"),
-                        Episode(description = "Para o Conselho Mundial. Princesa Vivi e Princesa Shirahoshi.",
-                                number = 777,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/777"),
-                        Episode(description = "Para o Conselho Mundial. Rebecca e o Reino das Sakuras.",
-                                number = 778,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/778"),
-                        Episode(description = "Kaidou Retorna. A Ameaça para a Pior Geração se Aproxima!",
-                                number = 779,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/779"),
-                        Episode(description = "A Frente de Batalha Faminta. Luffy e os Supernovas da Marinha!",
-                                number = 780,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/780"),
-                        Episode(description = "Os Três Implacáveis. A Grande Caçada aos Chapéus de Palha!",
-                                number = 781,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/781"),
-                        Episode(description = "O Punho do Demônio. A Luta Final! Luffy Vs Grant.",
-                                number = 782,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/782"),
-                        Episode(description = "Sanji Volta para Casa. Rumo ao Território da Big Mom!",
-                                number = 783,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/783"),
-                        Episode(description = "0 e 4. Encontro Inesperado! Germa 66.",
-                                number = 784,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/784"),
-                        Episode(description = "A Crise do Veneno Mortal. Luffy e Reiju!",
-                                number = 785,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/785"),
-                        Episode(description = "Totto Land! A Yonkou Big Mom Aparece.",
-                                number = 786,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/786"),
-                        Episode(description = "A Filha da Yonkou. A Noiva do Sanji, Purin.",
-                                number = 787,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/787"),
-                        Episode(description = "O Ataque Titânico! O Sofrimento de Fome da Mãe.",
-                                number = 788,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/788"),
-                        Episode(description = "A Capital em Colapso?! Big Mom e Jinbe.",
-                                number = 789,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/789"),
-                        Episode(description = "O Castelo da Yonkou. A Chegada na Ilha Whole Cake.",
-                                number = 790,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/790"),
-                        Episode(description = "Uma Floresta de Doces. Luffy Vs Luffy?!",
-                                number = 791,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/791"),
-                        Episode(description = "Os Assassinos da Mama. Luffy e o Bosque da Sedução!",
-                                number = 792,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/792"),
-                        Episode(description = "Judge.",
-                                number = 793,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/793"),
-                        Episode(description = "Confronto de Pai e Filho. Judge VS Sanji!",
-                                number = 794,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/794"),
-                        Episode(description = "Uma Ambição Gigantesca. Big Mom e Caesar.",
-                                number = 795,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/795"),
-                        Episode(description = "O País das Almas. A Temível Habilidade da Mãe!",
-                                number = 796,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/796"),
-                        Episode(description = "O Grande Comandante. Um dos Três Generais, Cracker, Aparece.",
-                                number = 797,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/797"),
-                        Episode(description = "O Inimigo de 800 Milhões. Luffy VS Cracker dos Mil Braços.",
-                                number = 798,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/798"),
-                        Episode(description = "Confronto de Forças Máximas. Gear 4 VS Habilidade da Bisu Bisu.",
-                                number = 799,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/799"),
-                        Episode(description = "1 e 2. Reunião! A Família Vinsmoke.",
-                                number = 800,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/800"),
-                        Episode(description = "O Salvador de uma Vida. Sanji e Chef Zeff.",
-                                number = 801,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/801"),
-                        Episode(description = "A Fúria de Sanji. O Segredo da Germa 66.",
-                                number = 802,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/802"),
-                        Episode(description = "Abandonando o Passado. Vinsmoke Sanji.",
-                                number = 803,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/803"),
-                        Episode(description = "Para o East Blue. A Decisão da Partida de Sanji.",
-                                number = 804,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/804"),
-                        Episode(description = "A Batalha de Limites. Luffy e os Biscoitos Infinitos.",
-                                number = 805,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/805"),
-                        Episode(description = "Tankman!",
-                                number = 806,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/806"),
-                        Episode(description = "Uma Luta Triste. Luffy Vs Sanji. (Parte 1)",
-                                number = 807,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/807"),
-                        Episode(description = "Uma Luta Triste. Luffy Vs Sanji. (Parte 2)",
-                                number = 808,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/808"),
-                        Episode(description = "Uma Tempestade de Vingança. O Ataque do Exército da Fúria!",
-                                number = 809,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/809"),
-                        Episode(description = "O Fim da Aventura. Sanji Escolhe a Proposta de Casamento.",
-                                number = 810,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/810"),
-                        Episode(description = "Vou Esperá-lo Aqui. Luffy Vs Exército da Fúria.",
-                                number = 811,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/811"),
-                        Episode(description = "Invadindo o Chateau. Roubem! O Road Poneglyph.",
-                                number = 812,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/812"),
-                        Episode(description = "Encontro Predestinado. Luffy e Big Mom!",
-                                number = 813,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/813"),
-                        Episode(description = "O Grito da Alma. O Ataque Relâmpago Estratégico de Brook e Pedro.",
-                                number = 814,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/814"),
-                        Episode(description = "Adeus. As Lágrimas da Decisão da Purin.",
-                                number = 815,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/815"),
-                        Episode(description = "A Fatalidade dos Olhos Esquerdos. Pedro Vs Barão Tamago.",
-                                number = 816,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/816"),
-                        Episode(description = "Cigarro Molhado. Sanji e a Véspera do Casamento.",
-                                number = 817,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/817"),
-                        Episode(description = "A Alma Indomável. Brook Vs Big Mom.",
-                                number = 818,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/818"),
-                        Episode(description = "O Desejo de Sora. Sanji, o Fracasso da Germa.",
-                                number = 819,
-                                animeName = "One Piece",
-                                link = "https://one-piece-x.com.br/episodios/online/819")
-                )
-        )
-    }
-
-    init {
-        UrlFetcher.useCache = true
-        FactoryChecker.checkFactory(OnePieceXFactory, VALID_URLS, INVALID_URLS, currentEpisode)
-    }
-
-}
