+++ settings.gradle
@@ -5,4 +5,3 @@ include ':cli'
 include ':decoders:animesorion'
 include ':decoders:animesproject'
 include ':decoders:anitubex'
-include ':decoders:onepiecex'
